requirements.txt
aiogram>=3.13.0
aiocryptopay>=0.4.8
aiohttp>=3.9.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
